Packages geupdate naar de laatste versie. Unittest framework versimpeld. Referenties in projecten opgeschoond.
