+++ PowerpointGenerator/Screens/LiturgieEdit.cs
@@ -1,6 +1,5 @@
 ﻿using ILiturgieDatabase;
 using System;
-using System.Data;
 using System.Linq;
 using System.Windows.Forms;
