+++ PowerpointGenerator/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Autofac" version="4.5.0" targetFramework="net46" />
-  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net46" />
+  <package id="Autofac" version="4.6.2" targetFramework="net46" />
+  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net46" />
 </packages>
