+++ RemoteGenerator/packages.config
@@ -1,20 +1,20 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Autofac" version="4.5.0" targetFramework="net46" />
+  <package id="Autofac" version="4.6.2" targetFramework="net46" />
   <package id="Autofac.Configuration" version="4.0.1" targetFramework="net46" />
-  <package id="CommonServiceLocator" version="1.3" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration" version="1.1.1" targetFramework="net46" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="1.1.1" targetFramework="net46" />
-  <package id="Microsoft.Extensions.Primitives" version="1.1.0" targetFramework="net46" />
-  <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="net46" />
+  <package id="CommonServiceLocator" version="2.0.3" targetFramework="net46" />
+  <package id="Microsoft.Extensions.Configuration" version="2.0.0" targetFramework="net461" />
+  <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.0.0" targetFramework="net461" />
+  <package id="Microsoft.Extensions.Primitives" version="2.0.0" targetFramework="net461" />
+  <package id="Microsoft.NETCore.Platforms" version="2.0.1" targetFramework="net461" />
   <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="net46" />
-  <package id="NETStandard.Library" version="1.6.1" targetFramework="net46" />
+  <package id="NETStandard.Library" version="2.0.1" targetFramework="net461" />
   <package id="System.AppContext" version="4.3.0" targetFramework="net46" />
   <package id="System.Collections" version="4.3.0" targetFramework="net46" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net46" />
   <package id="System.Console" version="4.3.0" targetFramework="net46" />
   <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="net46" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.3.0" targetFramework="net46" />
+  <package id="System.Diagnostics.DiagnosticSource" version="4.4.1" targetFramework="net461" />
   <package id="System.Diagnostics.Tools" version="4.3.0" targetFramework="net46" />
   <package id="System.Diagnostics.Tracing" version="4.3.0" targetFramework="net46" />
   <package id="System.Globalization" version="4.3.0" targetFramework="net46" />
@@ -26,7 +26,7 @@
   <package id="System.IO.FileSystem.Primitives" version="4.3.0" targetFramework="net46" />
   <package id="System.Linq" version="4.3.0" targetFramework="net46" />
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net46" />
-  <package id="System.Net.Http" version="4.3.1" targetFramework="net46" />
+  <package id="System.Net.Http" version="4.3.3" targetFramework="net461" />
   <package id="System.Net.Primitives" version="4.3.0" targetFramework="net46" />
   <package id="System.Net.Sockets" version="4.3.0" targetFramework="net46" />
   <package id="System.ObjectModel" version="4.3.0" targetFramework="net46" />
@@ -35,22 +35,22 @@
   <package id="System.Reflection.Primitives" version="4.3.0" targetFramework="net46" />
   <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="net46" />
   <package id="System.Runtime" version="4.3.0" targetFramework="net46" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.3.0" targetFramework="net46" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0" targetFramework="net461" />
   <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net46" />
   <package id="System.Runtime.Handles" version="4.3.0" targetFramework="net46" />
   <package id="System.Runtime.InteropServices" version="4.3.0" targetFramework="net46" />
   <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net46" />
   <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="net46" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Algorithms" version="4.3.1" targetFramework="net461" />
   <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net46" />
   <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net46" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.X509Certificates" version="4.3.2" targetFramework="net461" />
   <package id="System.Text.Encoding" version="4.3.0" targetFramework="net46" />
   <package id="System.Text.Encoding.Extensions" version="4.3.0" targetFramework="net46" />
   <package id="System.Text.RegularExpressions" version="4.3.0" targetFramework="net46" />
   <package id="System.Threading" version="4.3.0" targetFramework="net46" />
   <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net46" />
   <package id="System.Threading.Timer" version="4.3.0" targetFramework="net46" />
-  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="net46" />
+  <package id="System.Xml.ReaderWriter" version="4.3.1" targetFramework="net461" />
   <package id="System.Xml.XDocument" version="4.3.0" targetFramework="net46" />
 </packages>
