+++ Tests/PowerpointGenerator.Tests/LiturgieOplosserTests.cs
@@ -1,28 +1,28 @@
-﻿// Copyright 2016 door Erik de Roos
+﻿// Copyright 2018 door Erik de Roos
 using FakeItEasy;
-using ILiturgieDatabase;
-using NUnit.Framework;
 using Generator.Database.FileSystem;
+using ILiturgieDatabase;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Generator.Tests
 {
     public class LiturgieOplosserTests
     {
-        private static string DefaultEmptyName = "!leeg";
+        private const string DefaultEmptyName = "!leeg";
         private ILiturgieInterpreteer _liturgieInterpreteer;
 
-        [OneTimeSetUp]
+        [TestInitialize]
         public void Initialise()
         {
             _liturgieInterpreteer = A.Fake<ILiturgieInterpreteer>();
         }
 
-        [TestFixture]
+        [TestClass]
         public class LosOpMethod : LiturgieOplosserTests
         {
-            [TestCase("Psalm", "100")]
+            [DataTestMethod]
+            [DataRow("Psalm", "100")]
             public void NormaalItem_Gevonden(string onderdeel, string fragment)
             {
                 var liturgieItem = FakeInterpretatie(onderdeel, fragment: fragment);
@@ -31,10 +31,11 @@ namespace Generator.Tests
 
                 var oplossing = sut.LosOp(liturgieItem);
 
-                Assert.That(oplossing.Resultaat, Is.EqualTo(LiturgieOplossingResultaat.Opgelost));
+                Assert.AreEqual(oplossing.Resultaat, LiturgieOplossingResultaat.Opgelost);
             }
 
-            [TestCase("Welkom")]
+            [DataTestMethod]
+            [DataRow("Welkom")]
             public void CommonItem_Gevonden(string onderdeel)
             {
                 var liturgieItem = FakeInterpretatie(onderdeel);
@@ -43,11 +44,12 @@ namespace Generator.Tests
 
                 var oplossing = sut.LosOp(liturgieItem);
 
-                Assert.That(oplossing.Resultaat, Is.EqualTo(LiturgieOplossingResultaat.Opgelost));
+                Assert.AreEqual(oplossing.Resultaat, LiturgieOplossingResultaat.Opgelost);
             }
 
-            [TestCase("Psalm", "100")]
-            [TestCase("Welkom", null)]
+            [DataTestMethod]
+            [DataRow("Psalm", "100")]
+            [DataRow("Welkom", null)]
             public void AlleItems_GebruikStandaardNaam(string onderdeel, string fragment)
             {
                 var liturgieItem = FakeInterpretatie(onderdeel, fragment: fragment);
@@ -56,10 +58,11 @@ namespace Generator.Tests
 
                 var oplossing = sut.LosOp(liturgieItem);
 
-                Assert.That(oplossing.Regel.Display.Naam, Is.EqualTo(onderdeel));
+                Assert.AreEqual(oplossing.Regel.Display.Naam, onderdeel);
             }
 
-            [TestCase("Psalm2", "100", "Psalm")]
+            [DataTestMethod]
+            [DataRow("Psalm2", "100", "Psalm")]
             public void NormaalItem_GebruikDisplay(string onderdeel, string fragment, string display)
             {
                 var liturgieItem = FakeInterpretatie(onderdeel, fragment: fragment);
@@ -68,10 +71,11 @@ namespace Generator.Tests
 
                 var oplossing = sut.LosOp(liturgieItem);
 
-                Assert.That(oplossing.Regel.Display.Naam, Is.EqualTo(display));
+                Assert.AreEqual(oplossing.Regel.Display.Naam, display);
             }
 
-            [TestCase("Psalm2", "100", "psalm2", "Psalm")]
+            [DataTestMethod]
+            [DataRow("Psalm2", "100", "psalm2", "Psalm")]
             public void NormaalItem_GebruikMask(string onderdeel, string fragment, string maskRealName, string maskUseName)
             {
                 var maskList = FakeMask(maskRealName, maskUseName);
@@ -81,10 +85,11 @@ namespace Generator.Tests
 
                 var oplossing = sut.LosOp(liturgieItem, maskList);
 
-                Assert.That(oplossing.Regel.Display.Naam, Is.EqualTo(maskUseName));
+                Assert.AreEqual(oplossing.Regel.Display.Naam, maskUseName);
             }
 
-            [TestCase("Welkom_groot", "Welkom_groot", "Welkom")]
+            [DataTestMethod]
+            [DataRow("Welkom_groot", "Welkom_groot", "Welkom")]
             public void CommonItem_GebruikMask(string onderdeel, string maskRealName, string maskUseName)
             {
                 var maskList = FakeMask(maskRealName, maskUseName);
@@ -94,7 +99,7 @@ namespace Generator.Tests
 
                 var oplossing = sut.LosOp(liturgieItem, maskList);
 
-                Assert.That(oplossing.Regel.Display.Naam, Is.EqualTo(maskUseName));
+                Assert.AreEqual(oplossing.Regel.Display.Naam, maskUseName);
             }
         }
 
@@ -112,76 +117,6 @@ namespace Generator.Tests
             A.CallTo(() => liturgieItem.Benaming).Returns(onderdeel);
             if (fragment != null)
                 A.CallTo(() => liturgieItem.Deel).Returns(fragment);
-            return liturgieItem;
-        }
-
-        private static ILiturgieDatabase.ILiturgieDatabase FakeDatabase(string onderdeel, string fragment, string display = null, LiturgieOplossingResultaat status = LiturgieOplossingResultaat.Opgelost)
-        {
-            var zoekresultaat = A.Fake<IOplossing>();
-            A.CallTo(() => zoekresultaat.OnderdeelNaam).Returns(onderdeel);
-            A.CallTo(() => zoekresultaat.FragmentNaam).Returns(fragment);
-            A.CallTo(() => zoekresultaat.Status).Returns(status);
-            if (display != null)
-                A.CallTo(() => zoekresultaat.OnderdeelDisplayNaam).Returns(display);
-            var database = A.Fake<ILiturgieDatabase.ILiturgieDatabase>();
-            A.CallTo(database)
-                .Where(d => d.Method.Name == "ZoekOnderdeel")
-                .WithReturnType<IOplossing>()
-                .WithAnyArguments()
-                .Returns(zoekresultaat);
-            return database;
-        }
-    }
-
-    public class LiturgieOplosser_Bijbeltekst
-    {
-        private static string DefaultEmptyName = "!leeg";
-        private ILiturgieInterpreteer _liturgieInterpreteer;
-
-        [OneTimeSetUp]
-        public void Initialise()
-        {
-            _liturgieInterpreteer = A.Fake<ILiturgieInterpreteer>();
-        }
-
-        [TestFixture]
-        public class LosOpMethod : LiturgieOplosser_Bijbeltekst
-        {
-
-            [TestCase("Johannes", "3", new[] { "1", "2", "3" })]
-            public void NormaalItem_ZoekInDatabase(string onderdeel, string fragment, IEnumerable<string> fragmentVerzen)
-            {
-                var liturgieItem = FakeBijbeltekstInterpretatie(onderdeel, fragment, fragmentVerzen);
-                var database = FakeDatabase(onderdeel, fragment);
-                var sut = (new Generator.LiturgieOplosser.LiturgieOplosser(database, _liturgieInterpreteer, DefaultEmptyName)) as ILiturgieLosOp;
-
-                var oplossing = sut.LosOp(liturgieItem);
-
-                A.CallTo(() => database.ZoekOnderdeel(VerwerkingType.bijbeltekst, liturgieItem.Benaming, liturgieItem.PerDeelVersen.First().Deel, liturgieItem.PerDeelVersen.First().Verzen)).MustHaveHappened();
-            }
-
-            [TestCase("Johannes", "3", new[] { "1", "2", "3" })]
-            public void NormaalItem_Gevonden(string onderdeel, string fragment, IEnumerable<string> fragmentVerzen)
-            {
-                var liturgieItem = FakeBijbeltekstInterpretatie(onderdeel, fragment, fragmentVerzen);
-                var database = FakeDatabase(onderdeel, fragment);
-                var sut = (new Generator.LiturgieOplosser.LiturgieOplosser(database, _liturgieInterpreteer, DefaultEmptyName)) as ILiturgieLosOp;
-
-                var oplossing = sut.LosOp(liturgieItem);
-
-                Assert.That(oplossing.Resultaat, Is.EqualTo(LiturgieOplossingResultaat.Opgelost));
-            }
-        }
-
-        private static ILiturgieInterpretatieBijbeltekst FakeBijbeltekstInterpretatie(string onderdeel, string fragment, IEnumerable<string> verzen)
-        {
-            var liturgieItem = A.Fake<ILiturgieInterpretatieBijbeltekst>();
-            A.CallTo(() => liturgieItem.Benaming).Returns(onderdeel);
-            A.CallTo(() => liturgieItem.Deel).Returns(fragment);
-            var liturgieItemDeel = A.Fake<ILiturgieInterpretatieBijbeltekstDeel>();
-            A.CallTo(() => liturgieItemDeel.Deel).Returns(fragment);
-            A.CallTo(() => liturgieItemDeel.Verzen).Returns(verzen.ToList());
-            A.CallTo(() => liturgieItem.PerDeelVersen).Returns(new[] { liturgieItemDeel });
             return liturgieItem;
         }
