+++ Tests/PowerpointGenerator.Tests/packages.config
@@ -1,8 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Autofac" version="4.5.0" targetFramework="net461" />
-  <package id="Autofac.Extras.FakeItEasy" version="4.0.0" targetFramework="net461" />
-  <package id="FakeItEasy" version="2.3.3" targetFramework="net461" />
-  <package id="NUnit" version="3.6.1" targetFramework="net461" />
-  <package id="NUnit3TestAdapter" version="3.7.0" targetFramework="net461" />
+  <package id="FakeItEasy" version="4.5.1" targetFramework="net461" />
+  <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net461" />
+  <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net461" />
 </packages>
