+++ Tools/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Autofac" version="4.5.0" targetFramework="net45" />
+  <package id="Autofac" version="4.6.2" targetFramework="net45" />
 </packages>
